feat: rename mggitverifynodiff to mggit
Shorter name with less repetition.
